backend/pyproject.toml
[project]
name = "bikepacker-tracker-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "jacantwell",email = "jasper66018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.1,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "boto3 (>=1.37.37,<2.0.0)",
    "botocore (>=1.37.37,<2.0.0)",
    "python-frontmatter (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
